deploy: 136a2d1de4e6c4b910c32bcd54143a98ff3bd02b

+++ RGFW.h
@@ -7220,7 +7220,7 @@ HICON RGFW_loadHandleImage(u8* src, i32 c, RGFW_area a, BOOL icon) {
     
     for (size_t y = 0; y < a.h; y++) {
         for (size_t x = 0; x < a.w; x++) {
-			u32 index = (y * 4 * a.w) + x * channels;
+			size_t index = (y * 4 * (size_t)a.w) + x * channels;
             target[index] = src[index + 2];
             target[index + 1] = src[index + 1];
             target[index + 2] = src[index];
